refactor(config): extract secondary-display integration out of SqliteConfigProvider
Move the native secondary-display / platform-channel cluster (9 methods)
verbatim into a new part `sqlite_config_provider/secondary_display.dart` as
`extension SqliteConfigSecondaryDisplay on SqliteConfigProvider`:
updateHideBottomScreen, _handleSecondaryDisplayCall, _onSecondaryDisplayChanged,
_onSecondaryStateChanged, _handleSecondaryScreenshotRequest,
resetSecondaryInGameState, pushScreenshotAccess, refreshSecondaryScreenshotAccess,
reapplySecondaryDisplay.

Same (a)-split kit as the mutators/scanning parts: public extension,
notifyListeners() routed through the host `_notify()` bridge, and the statics
`_log`/`_secondaryDisplayChannel` host-qualified (all provably equivalent). Host
keeps the class declaration, all state, lifecycle wiring, and helpers. Host diff
is a single `part` directive plus pure deletions (remaining host byte-identical).

Closes the H decomposition (mutators/scanning/secondary_display) and the whole
9-file refactor campaign. No behaviour or public-API change.
